Pipeline - Restore old $match functionality as Mongo 3.6.4 and then fallback to new $expr expressions. Mongo 3.6.4 server attempts to rewrite some or all of the $expr expression using the old $match operators so that it will hit indexes as appropriate.  In my real world test code, the following query is failing to hit the our index on Date using $expr.
    .Where(c => c.Date >= new DateTime(2018,6,1))
    { "$match" : { "$expr" : { "$gt" : ["$d", ISODate("2018-06-01T07:00:00Z")] } } }

Unfortunately, this puts the burden on the end user of this library to know what will trigger usage of $expr
and what wont.  For example, this would fall back to $expr:

    .Where(c => c.Date >= new DateTime(2018,6,1) && c.ArrayProp.All(d => d > 0))

This could be rewritten to multiple where calls so that only the latter falls back:
    .Where(c => c.Date >= new DateTime(2018,6,1))
    .Where(c => c.ArrayProp.All(d => d > 0))

A future enhancement here is for MongoLinqPlusPlus to split a Where predicate on
the boolean ANDs into multiple $match statements so that as much of the Where
predicate uses the old Match semantics as possible.

Until this happens - or Mongo server fixes this bug, we have a Unit test risk.
The unit tests should each target the old and new $match functionality so that
we know it all works.  Right now that isn't happening.